Improve user feedback for missing plugins #787
 * also ensure that the extractcode cli does not depend
   on scancode cli.
